UserlandFS: Make 64 bit clean and enable build for x86_64.
I did not bother to fix the BeOS kernel emulation, so this part is
still left out of the x86_64 build.

+++ src/add-ons/kernel/file_systems/userlandfs/kernel_add_on/FileSystem.h
@@ -81,7 +81,7 @@ public:
 			RequestPortPool*	GetPortPool();
 
 			status_t			Mount(fs_volume* fsVolume, const char* device,
-									ulong flags, const char* parameters,
+									uint32 flags, const char* parameters,
 									Volume** volume);
 //			status_t		 	Initialize(const char* deviceName,
 //									const char* parameters, size_t len);
